requirements.txt
GitPython >= 1.0.1
MySQL-python >= 1.2.5
PyYAML >= 3.10
boto >= 2.28.0
inflect >= 0.2.4
netaddr >= 0.7.11
requests >= 0.14.0
toposort >= 1.0
